Moving ServerLogic completely out of ServerShell
