+++ C#/BluffinMuffin.Protocol.Server/BluffinMuffin.Protocol.Server.csproj
@@ -57,14 +57,15 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Annotations\Annotations.cs" />
-    <Compile Include="BluffinGameWorker.cs" />
-    <Compile Include="BluffinLobbyWorker.cs" />
-    <Compile Include="BluffinTcpServer.cs" />
-    <Compile Include="GameCommandEntry.cs" />
-    <Compile Include="CommandEntry.cs" />
-    <Compile Include="IBluffinClient.cs" />
-    <Compile Include="IBluffinLobby.cs" />
-    <Compile Include="IBluffinServer.cs" />
+    <Compile Include="Workers\BluffinGameWorker.cs" />
+    <Compile Include="Workers\BluffinLobbyWorker.cs" />
+    <Compile Include="BluffinServer.cs" />
+    <Compile Include="LocalTcpServer.cs" />
+    <Compile Include="DataTypes\GameCommandEntry.cs" />
+    <Compile Include="DataTypes\CommandEntry.cs" />
+    <Compile Include="DataTypes\IBluffinClient.cs" />
+    <Compile Include="DataTypes\IBluffinLobby.cs" />
+    <Compile Include="DataTypes\IBluffinServer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RemotePlayer.cs" />
     <Compile Include="RemoteTcpClient.cs" />
