+++ C#/BluffinMuffin.Server/BluffinMuffin.Server.csproj
@@ -51,7 +51,6 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="BluffinServerLobby.cs" />
     <Compile Include="Program.cs">
       <SubType>
       </SubType>
